fixed a couple g++ compile errors, and changed the size_t in the gm header back to an unsigned int because it causes a template build error.

+++ trunk/delta3d/inc/dtGame/gamemanager.h
@@ -463,7 +463,7 @@ namespace dtGame
           * manager.
           * @return The number of game actors in the system.
           */
-         size_t GetNumGameActors() const { return mGameActorProxyMap.size(); }
+         unsigned GetNumGameActors() const { return mGameActorProxyMap.size(); }
         
          /**
           * Retrieves all the game actors added to the GM
